requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
SQLAlchemy==2.0.20
alembic==1.11.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0 